Move SunriseSunsetAPI entities to the SunriseSunset folder
