+++ app/src/main/java/com/korostenskyi/owlyweather/data/network/OpenWeatherDataSourceImpl.kt
@@ -1,0 +1,10 @@
+package com.korostenskyi.owlyweather.data.network
+
+import androidx.lifecycle.LiveData
+import com.korostenskyi.owlyweather.data.network.entity.OpenWeather.CurrentWeather
+
+class OpenWeatherDataSourceImpl : OpenWeatherDataSource {
+    override val fetchedCurrentWeather: LiveData<CurrentWeather>
+        get() =
+                
+}
