+++ app/src/main/java/com/korostenskyi/owlyweather/data/network/api/OpenWeatherApiClient.kt
@@ -1,0 +1,4 @@
+package com.korostenskyi.owlyweather.data.network.api
+
+interface OpenWeatherApiClient {
+}
