--- app/src/main/java/com/korostenskyi/owlyweather/data/network/entity/SunSet.kt
@@ -1,6 +1,0 @@
-package com.korostenskyi.owlyweather.data.network.entity
-
-data class SunSet(
-    val results: Results,
-    val status: String
-)
