+++ app/src/main/java/com/korostenskyi/owlyweather/data/network/entity/SunriseSunset/SunSet.kt
@@ -1,0 +1,6 @@
+package com.korostenskyi.owlyweather.data.network.entity.SunriseSunset
+
+data class SunSet(
+    val results: Results,
+    val status: String
+)
